Adjust the order of topic events notified to the p2pserver when the node starts

+++ consensus/cbft/consensus_process_rg_test.go
@@ -501,7 +501,7 @@ func TestRGViewChangeQuorumCert(t *testing.T) {
 	msg := &protocols.RGViewChangeQuorumCert{
 		GroupID:        groupID,
 		ViewChangeQC:   rgqc,
-		ValidatorIndex: uint32(validatorIndex),
+		ValidatorIndex: validatorIndex,
 		PrepareQCs: &ctypes.PrepareQCs{
 			QCs: []*ctypes.QuorumCert{qc},
 		},
